front_end/requirements.txt
google-cloud-secret-manager
pandas >= 1.2.0
numpy >= 1.20.0
scipy >= 1.6.0
websockets
alpaca-trade-api 
streamlit
sqlalchemy
google-cloud-storage
pyodbc
plotly
yfinance